Add rngseed for blocks.
Since the mining change, it is trivial for miners to brute force the
actual block hashes to find a favourable one in case games base their
random numbers on block hashes (because miners can do so before doing
the actual PoW).

To fix this, random numbers should be based off the actual PoW, i.e.
the hash of the fake header or the auxpow parent block.  To simplify
this, we add a standardised routine computing a hash based off the PoW
to the core daemon.  It is returned in the 'rngseed' field for getblock.

Note that miners can still, after doing the PoW, decide to withhold a
block whose resulting random numbers they do not like.  But this has
at least some cost (the block reward), so is only a concern for
high-stakes decisions in games.