Split out test-only code from auxpow.py module.
auxpow.py in test/functional/test_framework contains various utility
methods for dealing with auxpow's.  Originally created to support the
auxpow-related stuff in regtests, some common parts of it are now also
used for the getwork-wrapper.py script.

This cleanup splits the existing module into two parts:  The new
auxpow.py contains the common code that is also used for the getwork
wrapper (mainly the construction and handling of auxpow's), while the
test-only stuff (solving of auxpow's and utility methods for
verification) is moved to auxpow_testing.py.  This lets us better
separate the code.