Add 'height' field to name RPC outputs.
While there is no expiration of names in Xaya, names still have a
defined "last updated" height (that is actually even tracked internally
by the name database).  This change adds back the "addExpirationInfo"
function from upstream Namecoin, renamed as "addHeightInfo".  We use it
to attach at least the height field to the output of name RPCs,
as is done for upstream Namecoin (together with the expiration info).

This implements https://github.com/xaya/xaya/issues/47.