Reduce syncing wait time for name_multisig.
This changes name_multisig.py to spend less time syncing the nodes,
while keeping the basic functionality of the test in tact.  The
total number of test nodes is reduced to two, which is sufficient for
testing a 2-of-2 multisig.  We now also avoid syncing before and after
each block generation; instead, we only sync the blocks afterwards
at the places where it is actually necessary.

This speeds up the test from 20s to 2.3s on my machine.