Expand name vs P2SH regtests.
As discussed in https://github.com/namecoin/namecoin-core/issues/241,
this expands the regtests we have to ensure that name prefixes interact
correctly with P2SH outputs and redeem scripts.

In particular, we test that name prefixes in redeem scripts are just
ignored, while a name prefix for the output script itself (before the
P2SH scriptPubKey) works as expected.