Re-enable feature_block.py test.
The upstream test feature_block.py is disabled in Namecoin, as it
violates the BDB lock limit.  This is not present in Xaya, so we can
re-enable the test.  This only needs to update it to the reduced sigop
limit, and for the powdata change (nBits are now there and not on the
block itself).