Fix disconnecting block 1 with transactions.
ApplyTxInUndo contains code that supports a legacy format of undo
data, which fails block disconnect if block 1 has more than the
coinbase transaction.  This is not relevant for Bitcoin or Namecoin,
whose block 1 is empty, but breaks for Xaya.  To fix this, we can
simply remove the legacy code, as Xaya starts immediately with the
newer undo data format and so the support for legacy data is not
relevant for us.

This fixes https://github.com/xaya/xaya/issues/49.