New contrib script for bech32 address conversion.
This adds a new script in contrib/namecoin (convertBechAddress.py and a
helper convertBechAddress.sh wrapper).  The new script complements
convertAddress.py for bech32, i.e. it allows to convert an arbitrary
bech32 address (with some HRP) to another address with the same payload
but a different HRP.

This will be useful for updating test data from Bitcoin's to Namecoin's
address format.