Add lock for ZMQ publishers.
With Xaya's game_sendupdates, ZMQ notifications may be sent from
multiple threads at the same time.  Because furthermore sockets can
be shared between notifiers, we have to synchronise all publications
for the whole ZMQ subsystem.

(The ZMQ manual recommends to simply not share sockets among threads,
but that is not possible here as we cannot bind per-thread sockets to
the same publication address.  It is explicitly fine to share sockets
with proper locking, though.)