Update the signmessage "magic string".
signmessge and verifymessage use a "magic string" prefix.  Previously,
Namecoin was using the same as upstream Bitcoin.  This has the confusing
effect that signatures can be interchanged between Bitcoin and Namecoin.
(But of course, someone still has to own the private key to create a
valid signature for any of them in the first place.)

This commit changes the magic string to be specific to Namecoin, which
makes sure that messages signed in Namecoin can only be validated in
Namecoin, and messages signed in Bitcoin only with Bitcoin (as most
people likely expect).

Note that this means that old signatures won't validate anymore with
verifymessage.  But it is probably unlikely that this was used a lot
at all.  Furthermore, the signature can of course still be verified in
a custom way (e.g. using Bitcoin Core), so old signatures can still be
used as a proof if that is necessary.