Use configured encoding for "input" names/values.
Instead of using ValtypeFromString to convert names/values given in the
RPC interface to valtype, call DecodeName and take the configured
encoding into account.

(The input-side for the REST interface is using a custom function to
handle URL-escaped names and thus not related to this change.)

This implements all changes required for
https://github.com/namecoin/namecoin-core/issues/246 from the "input"
side.  The "output side" (e.g. "name" and "value" fields in the result
of name_show) is still missing.

With this change, all uses of ValtypeFromString have been removed except
for the one in DecodeName itself and the unit test.