Fix overflow in auxpow computation.
The expression (1 << h) can overflow, particularly if h is 31.  Using
(1u << h) avoids this for that particular case.

This is not really important in practice, as h > 30 is forbidden anyway by
separate code and thus the result for h >= 31 doesn't matter.  But it is
a condition triggered in the unit test and which fails with strict checks
that are turned on for the test binary.