Add missing lock for Misbehaving.
Misbehaving() expects cs_main to be held.  This was not the case for
the error raised when receiving an oversize headers message (which is
specific to auxpow and not in upstream Bitcoin).

This change should not affect performance, since the lock will only be
acquired when there is actually a mishaviour (and not for all header
syncing).

This fixes some more of the warnings reported in
https://github.com/namecoin/namecoin-core/issues/245.