Introduce EncodeNameForMessage for logs / errors.
The new function EncodeNameForMessage encodes a given name in an
encoding suitable for logs, errors and other messages (independent
of the chosen -nameencoding).

We use it to replace many of the existing ValtypeToString calls, where
the result is used for logs or errors.  We also use it for the name op
description returned from "gettransaction" and "listtransactions" for
wallet transactions, where it is part of a longer descriptive string.