Explicitly use UTF-8 for name_filter regexp check.
Instead of the soon-to-be-removed ValtypeToString, use DecodeName with
an explicit UTF-8 encoding for the regexp check in name_filter.  Names
that do not encode as UTF-8 are assumed to not match the check.

This also adds a test case with invalid UTF-8 to name_scanning.py to
make sure it works, and changes the test to use only one node and no
synchronisation to make it faster.