Handle encodings on the output side.
Use the configured name/value encodings also for return values of RPCs
(and the /name/ REST handler in JSON format).

This mostly finishes the implementation of the proposal in
https://github.com/namecoin/namecoin-core/issues/246, except for making
the encoding configurable per-RPC through an options argument.

In particular, we now use the encoding options when populating the
'nameOp' JSON object for decoded raw transcations and scripts, and for
returning data from name_show, name_scan and the likes.

Those JSON objects got new fields 'name_encoding' and 'value_encoding',
so that the 'name' and 'value' fields can be interpreted correctly.
When a name/value cannot be encoded for the chosen encoding, then instead
of setting 'name'/'value', we set 'name_error' or 'value_error' to
indicate this.