Support -name/valueencoding for namecoin-tx.
namecoin-tx uses the same logic as namecoind when converting scripts
and transactions with name operations to JSON (with the -json argument).
Thus, it can take advantage of encoding support in the resulting JSON.

This change adds -nameencoding and -valueencoding arguments to
namecoin-tx, which function similar to the arguments of namecoind for
JSON output.

Note that they *do not* alter the *input* format, which is always hex
for namecoin-tx!