Verify block hashes when building a chain.
While building a chain of blocks for attach/detach tests in
xaya_gameblocks, also verify that the block hashes in the chain relate
to each other as expected.

This adds an extra layer of safety on the test to ensure that things are
fine; there is already a basic test for the child and parent hashes in
an attach notification, but none for the detach notifications or in a
"larger" setting (with more than one block).