Separate method for decoding value from RPC.
Instead of having DecodeNameFromRPCOrThrow decode both names and values
for RPC inputs, this adds a new DecodeValueFromRPCOrThrow for values.
The interface of the two functions now accepts not the actual encoding,
but an "options" UniValue from which the encoding (if explicitly
specified) will be taken in the future.  By separating the functions for
names and values, we make the calls more explicit and also can take care
of using the right value (and default value) for names vs values in the
functions themselves.

For now, the new options object is not actually set to anything
non-trivial.  This will be changed in a follow-up where we implement
options for write-side RPCs.