Allow encoding override in write RPCs.
Extend the "options" RPC argument that is already present on name_new,
name_firstupdate and name_update to include "nameEncoding" and
"valueEncoding" as new fields.  They can be used to override the name
and value encodings for RPC arguments on a per-RPC basis.

This implements a further missing piece (but not yet all) of
https://github.com/namecoin/namecoin-core/issues/246.