Define variable for config file in regtests.
Instead of hardcoding "namecoin.conf" as the name of the config file
in various files throughout the regtest framework, define a variable for
it in util.py and use it everywhere.

This makes rebranding for other projects based on Namecoin (like Xaya or
Huntercoin) easier, and also makes the code more flexible in general.