Disable invalid tests from feature_block.py.
Some of the tests in feature_block.py are not valid in Namecoin, because
they either run into the BDB lock limit or because they test strict
enforcement of BIP30, which is not active in Namecoin at all.

This disables (removes) those tests to make feature_block.py pass again
after a recent upstream change that made the checks stricter.