Reduce work requirement for assumed-valid block.
Upstream Bitcoin requires that a block be confirmed for at least two
week before it can qualify for -assumevalid.  This is meant to make
it hard for an attacker to have an invalid block accepted by telling
users to set -assumevalid.

For Xaya, that means that many more blocks than in Bitcoin are required
to be on top of the assumed-valid block, making feature_assumevalid.py
fail currently (or rather, the test is ignored).

This change reduces the work requirement, so that the test works fine
in the upstream version.  This is a bit of a hack, since ideally the
test should be updated instead; but that is infeasible, given that
already the 2,100 blocks created now push the resource requirements.
Furthermore, the risk of such an attack is much smaller in Xaya, so
that this change is certainly fine.  Afterwards, one day of blocks is
required instead of one week.