Add -maxgameblockattaches option and limit.
The new option -maxgameblockattaches (default: 1000) can be used to
specify the maximum number of block attach steps that will be sent
in response to a single game_sendupdates request.  The RPC response
will always indicate the "toblock" correctly, i.e. give the block to
which notifications will be sent (after limiting).

This won't impact client software like libxayagame, as it will expect
notifications to the returned "toblock" and then just request more
updates if necessary.

With this limit, we avoid problems with very large requests (e.g. the
whole chain when a game is synced from scratch) that will run for
minutes and block other requests for notifications in the mean time.
With the limit, those requests will be interleaved as each individual
request can only take a small amount of processing.

This fixes https://github.com/xaya/xaya/issues/65.