Allow non-zero nonce with POST_ICO fork.
This implements a hard fork scheduled at certain heights (500 on regtest,
1M on testnet and mainnet for now, will be changed) that allows non-zero
nonce values in the main block header (not the fake header).  This solves
https://github.com/xaya/xaya/issues/50.

Forks are added to Consensus::Params similar to how they are handled in
Huntercoin.  For now, there's only this one, but there will likely be
more in the future; for instance, to change the standalone mining
algorithm or when we implement ephemeral timestamps.

The created POST_ICO fork will contain more things in the future, namely
a change to the block reward and block interval as described in
https://github.com/xaya/xaya/issues/40.