Make block target spacing fork-able.
Instead of having the nPowTargetSpacing a simple chain param, turn it
into a method on the chain rules so that it can depend both on the
algorithm and on the block height (so that it can be changed in forks).

The code using the value is separated into parts that use the real
target spacing per algo (mainly for difficulty retargeting) and code
that actually needs the *average* spacing among all algorithms for
estimates of tip staleness and related stuff.  The latter is now
achieved using a new AvgTargetSpacing method, that computes the value
for all algorithms from the pure algo-dependent values (but still
depends on the height in case the value changes through forks in the
future).

The new flexibility is used to set the target spacing to 40 seconds
with Neoscrypt and 120 seconds with SHA-256d merged mining.  This will
allow us to get the desired 25%/75% split of block rewards between the
algorithms, while keeping the average at two blocks per minute for now.

Note that the average target spacing is not changed (even though the
code is made to support changes with it), so that the block heights
passed to AvgTargetSpacing do not matter for now.  The only changes
are due to some places where the nPowTargetSpacing was used while
it should have been nPowTargetSpacing/2 instead.  (For instance,
the -assumevalid time window had to be decreased again for this
to make the test with 2,100 blocks as window pass.)

This implements https://github.com/xaya/xaya/issues/67.