Add options argument to getNameInfo.
Refactor the encoding handling in the name RPCs a bit; in particular,
add an options argument to all getNameInfo variants, that allows to
override the configured default encodings for names/values that is
applied for the output.

For now, NO_OPTIONS is passed at all call sites.  This will be changed
in a follow-up, though, to ultimately implement options overrides for
read-type RPCs.  This is a still missing piece for resolving
https://github.com/namecoin/namecoin-core/issues/246.