Move help builder for "options" to rpc/names.
Move the help-text builder object for the name RPC "options" argument
from wallet/rpcnames to rpc/names.  So far, it was only used for the
write RPCs in wallet/rpcnames.  In the future, however, it will also
be useful for read-only RPCs like name_show (namely for the encoding
override options).