Add filtering options for name_scan.
This extends the "options" argument for name_scan to allow filtering
of names by the number of confirmations, the prefix and a regexp as
described in https://github.com/namecoin/namecoin-core/issues/237.
("showExpired" is not yet implemented, as that will be part of a more
general change also for other read-only RPCs in a later step.)

With these options, name_scan now supports every feature of name_filter
except for the "stats" mode.  The latter will be removed in a follow-up
change; the name_scanning.py regtest is already updated to only test
name_scan (but all the new features) in this commit.

This also implements https://github.com/namecoin/namecoin-core/issues/16
through the new "prefix" option for name_scan.