Check in processed data for HUC snapshot.
The new data file was produced by the processSnapshot.py Python script
from the already checked-in raw input data in snapshot-balances.json.
It has been posted to various forum threads for the community to check,
and is now the finalised data that will be used for the actual snapshot
payments.