Generic override for -minrelaytxfee in tests.
Add a generic feature to the regtest framework that sets
-minrelaytxfee=0.00001 unless -minrelaytxfee is already explicitly set.
This allows to use the upstream default value for regtests, so that
hardcoded fees in tests do not have to be changed for Namecoin's higher
default relay fee.

For now, the new feature is applied to the two tests that previously
set -minrelaytxfee by themselves for this reason (feature_rbf.py and
feature_fee_estimation.py), as well as rpc_net.py where -minrelaytxfee
is set by the test (so that we can verify this still works).

In the future, this should be applied for all tests without having to
"opt in" through enabling use_bitcoin_relay_fees.  For now, though, this
allows to migrate tests one-by-one.

See https://github.com/namecoin/namecoin-core/issues/267.