Increase block reward with POST_ICO fork.
Increase the block reward on testnet and mainnet together with the other
changes in the POST_ICO fork.  That means that the reward will be 1 CHI
before the fork, and then increase to a currently-set value of 10 CHI
afterwards.  It will halve as usual from there on.

On regtest, this rule is disabled, so that the reward stays 50 CHI with
normal halving.  This ensures that no magic numbers in tests have to be
tweaked unnecessarily.

The value of 10 CHI for the block reward is arbitrary.  It will be used
on testnet for verifying the fork.  On mainnet, the number will be changed
once the final fork height and thus the correct block reward (to achieve
the targeted total supply) have been fixed.