Strip name prefix for IsWitnessProgram.
When the segwit-code was originally integrated with Namecoin, we
did not update CScript::IsWitnessProgram to be aware of name prefixes.
This has the effect that names sent to a pure segwit script (e.g. a
bech32 address) can be updated by anyone even after segwit is activated
for Namecoin.  P2SH-Segwit addresses are working already as expected.

In this change, we fix that bug.  With the fix, IsWitnessProgram strips
a name prefix if it exists before extracting the witness program, so
that pure segwit scripts can safely hold not only coins but also names
after segwit activates.

This also includes new regression and unit tests for names and segwit,
including (but not limited to) those suggested by Jeremy in
https://github.com/namecoin/namecoin-core/issues/258.