Implement game admin commands for ZMQ gameblocks.
This adds support for "game admin commands" as described in
https://github.com/xaya/Specs/commit/43d0e7bdc4b91c1bf760d7bca8c3943bf46bedf7
to Xaya Core's gameblocks ZMQ notifications.

In particular, new blocks are now also checked for updates to g/ names.
If any update is found that contains a "cmd" value, then that is included
in the notification's payload data.

This implements https://github.com/xaya/xaya/issues/73.