Set POST_ICO fork on mainnet.
This fixes the POST_ICO fork on mainnet to be at block 440k and the initial
block reward after the fork to 3.82934346 CHI on mainnet.  For more details,
see https://github.com/xaya/xaya/issues/70#issuecomment-441292533.

Also the unit test for the total subsidy is re-enabled to verify that the
resulting PoW coin supply matches the expected amount.