Remove "ismine" for RPC calls without wallet.
When no wallet is associated to an RPC call like name_show, do not include
"ismine" in the result at all.  Previously, "ismine=false" was returned
in those cases.  But this is misleading, as it implies that the name is
certainly *not* in the wallet.

Instead, we now do not include "ismine" as wallet-specific field at all.
This means that especially multi-wallet situations without a specified
wallet now behave in the same way as if the wallet were not compiled in
at all.  This makes the behaviour consistent.

Note that standard situations where the user has exactly one wallet are
not affected at all by this change.  Only multi-wallet users are.