Fix disk block index for non-zero nNonce values.
After the post-ICO fork, nNonce values in blocks are allowed to be
non-zero.  This is fine from a consensus point of view (and was tested),
but it caused corruption to the on-disk block index.

The reason for this is that the nNonce value was not included in the
C(Disk)BlockIndex instances, and was assumed to be zero when computing
the block hash from a CDiskBlockIndex that is being loaded.  Thus blocks
would be inserted into mapBlockIndex at a wrong hash, so that the pprev
pointer of the next block (referring to the correct block index) was
pointing to an empty struct instead.

This fix adds back nNonce to C(Disk)BlockIndex, so that all works fine.
Users who synced beyond the first affected block (529761 on mainnet).

See also https://github.com/xaya/xaya/issues/82.