Allow extracting address in verifymessage.
This modifies the verifymessage RPC so that it can (optionally) extract
the address from the signature and return it.  In particular, when the
address is passed as "", then the signature is used to recover the
public key and the corresponding address is returned (if the recovery is
successful).

See https://github.com/bitcoin/bitcoin/pull/3473.