Add auxpow to getblock help text.
The getblock help text was missing a description of the "auxpow" field
in the returned JSON object.  This adds the missing help text.

Note that the field itself was already present and is not changed.  This
is purely about the documentation in the online help.

Fixes https://github.com/namecoin/namecoin-core/issues/289.