Add test for valid block with invalid auxpow.
The new regtest verifies that it is "fine" to submit a valid block with
invalid auxpow, and that this does not permanently mark the block as
invalid.  In other words, the node will later still accept the block
if a valid auxpow is supplied.

If this were not the case, then this could be used to fork the network.
The behaviour of the main validation code is already good, but by adding
an explicit test we can ensure that we do not regress on this.