Zero CAuxPow's hashBlock field.
As noticed in https://github.com/namecoin/namecoin-core/issues/288,
the hashBlock field of the coinbase Merkle tx in a CAuxPow object is
never actually used for anything.  That is not necessary, since the
parent block is known anyway, being another part of the CAuxPow.

With this change, we always set the hashBlock field to zero when loading
a CAuxPow into memory, and verify it is zero when writing it back (to disk
or a network peer).  This does not break anything, as is also verified
with a new regtest.

Having the field set to zero has two advantages:  First, it makes auxpow
data better compressible, allowing for space savings e.g. for SPV clients.
Second, it prevents accidental usage of the field in the future.

This is a first step towards the full resolution sketched in
https://github.com/namecoin/namecoin-core/issues/288.  The next steps will
be changing the structure of CAuxPow completely, so that this and other
redundant fields are removed.  That, however, is not backwards compatible,
while this simple change is.