Decouple CAuxPow from CBaseMerkleTx.
This refactors CAuxPow to be independent of CBaseMerkleTx.  Instead,
the fields and serialisation code are put directly in CAuxPow as needed.
By doing so, we can also get rid of two redundant fields:  The hashBlock
(which is always zero anyway since the last commit) and nIndex, which is
also always zero for a coinbase tx.  Those fields are still present in
the serialisation format (for backwards compatibility), but not in the
actual instance.

Note that this changes the consensus behaviour:  Previously, CAuxPow
instances with a non-zero nIndex value would be invalid.  With this change,
the nIndex value is ignored, so that those will be accepted.  That is,
however, fine, since the auxpow is just used to provide PoW for a block
and does not otherwise influence the consensus state.  If an attacker
makes use of this change to send out blocks that are accepted by new
nodes while they are rejected by old nodes, this is equivalent to the
attacker simply refusing to send blocks to old nodes (which they can always
do if they wish, independent of that particular change).

For reference, see https://github.com/namecoin/namecoin-core/issues/288.