Test multiple updates of a name in one block.
Namecoin's memory pool implementation does not allow multiple updates
to a single name (in a chain), but the consensus logic is fine with
that.  In other words, it is possible to update a name multiple times
in a single block as long as those updates are in that one block.

This adds a regression test for this situation as well as the mempool
handling, to make sure that the current and expected state is preserved
and no accidental regression is introduced with this edge case.

Also includes a tiny change to the logic of the mempool acceptance itself,
so that a proper error code is returned if accepting a transaction fails
due to name rules.