Schedule segwit activation softfork.
This schedules a softfork activating BIP16 (P2SH), Segwit and
CSV at a future blockheight on mainnet and testnet.

On mainnet, the activation will happen at block 475,000, which is
expected to be mined in about half a year.  On testnet, the activation
happens at block 232,000, or in one or two weeks (but depending strongly
on the mining activity on testnet).

See https://github.com/namecoin/namecoin-core/issues/239 for a discussion,
including the commitment to update timely by many of the most important
mining pools.