Do not install libnamecoinconsensus.
With this change, we disable installation of libnamecoinconsensus
when "make install" is done for Xaya.  The library has not been
rebranded, so that this would conflict with an installation of
Namecoin on the same system.  Since the library is not used anywhere,
this is also no loss of functionality (but once it is needed, it can
be properly rebranded and installation re-enabled).

After this change, only the manpages are still conflicting with Namecoin
and Bitcoin during installation.  This is an upstream issue with
Namecoin, and will be fixed there first.