Allow multiple updates of a name per block.
This removes the mempool-restriction of allowing only one update
per name and block.  That restriction was introduced originally to
simplify the mempool implementation, but is actually easy to lift.

With this change, we allow a chain of updates of one name to be in the
mempool at once; this can be useful e.g. to quickly fix an erraneous
previous update without the need to wait for confirmations in between.
For name_new and name_firstupdate, no changes are made.  Only updates
are allowed to be chained on a pending operation, although updates can
be chained to a name_firstupdate.

Note that the consensus layer of Namecoin already allows such chains,
and the restriction lifted is purely a policy one (in other words, this
is not a fork of any sort).  For now, we only remove the restriction
from the mempool itself; the name_update RPC method still won't allow
creating such chains.  We will remove the RPC restriction in a second
step later on, once miners and nodes have updated to the new policy
so that such chains of updates can actually be expected to be relayed
and mined.

This fixes https://github.com/namecoin/namecoin-core/issues/304.