Allow multiple name updates per block.
This is a backport of 7e31ae32e635bf1090c90395f7d2e85b2039c100, which
essentially is https://github.com/namecoin/namecoin-core/pull/322,
for the Xaya 1.2 release branch.

This allows a name to be updated multiple times in the mempool (on the
consensus level, this is already allowed in blocks).  The name_update
RPC method still does not allow this feature, as we need miners and
relaying nodes to update first.

By backporting this, we can make sure that miners and relaying nodes
update as soon as possible, so that we can soon take advantage of this
feature in actual games.