Use RPCHelpMan::Check for auxpow RPCs.
Make use of the new RPCHelpMan::Check utility method in the auxpow
mining RPCs where applicable (createauxblock and submitauxblock).

getauxblock has a special rule of what arguments are acceptable (exactly
zero or two), which is not reflected by RPCHelpMan alone and thus
we cannot make use of Check for this RPC method.