Allow multiple name_update RPCs per block.
Since "recently", the mempool for Xaya Core allows multiple pending
updates of a single name.  Most of the network has updated as well
to 1.2.1 (with this change), so that those transactions will get relayed
and mined fine.

In this change, we remove the restriction put into the name_update RPC,
that was not yet allowing such chains of transactions to be created.
After that, we can now "fully" make use of multiple pending name_update's
per name and block.