Add tests for JSON values with duplicate keys.
JSON objects may have duplicated keys, and libunivalue accepts that.
Since we use JSON parsing as validation in Xaya (for name values), we
should make sure this is clearly documented and verified.  Thus this
change adds a unit and regtest that both make sure that a JSON object
with duplicated key is accepted as name value.