Expose median time to games.
This adds the "mediantime" field to ZMQ block notifications, which
contains the "median time" of the currently attached/detached block.
Exposing this field to games can be useful, since it is a measure of time
that is guaranteed to be non-decreasing over time.

Implements https://github.com/xaya/xaya/issues/98.