Option to limit the length of pending name ops.
This introduces a new config option, -limitnamechains.  It sets the
length of "pending name operation chains" that name_update will create.

For now, this defaults to one, meaning that no chains of pending
name operations will be allowed.  This generalises the current check
in the name_update RPC code, which is there to prevent those for now
until the network has upgraded to code that supports chains in the
mempool itself.

With this, we make it easier to enable pending chains once the network
is sufficiently updated (and also allow each user of Namecoin Core to
make the decision for themselves, if they wish).  Furthermore, by having
a limit even in the future, we prevent "stuck" transactions resulting
from long chains of name_update's when they exceed the general mempool
ancestor limit (as happened on Xaya for instance, see
https://github.com/xaya/xaya/issues/99).