Properly handle duplicated JSON keys.
Xaya Core does some non-trivial processing of JSON values, namely
when determining what ZMQ notifications to send out for blocks
and pending transactions.

JSON allows duplicated keys in objects, which is an edge case that has
not been considered very well before.  With this update, we explicitly
handle it and specify what behaviour Xaya Core should have in tests
and the documentation.

In particular, if a name_update has duplicate "g" or "cmd" fields, then
we now process all of them.  If a single game ID is given multiple times
(because of duplicated "g" fields, or because it is a duplicated key
itself within a "g"), then we dedup that and only keep the last value.

Only sending one move for a single game (rather than multiple ones)
makes sense, because this avoids a situation where e.g. a single CHI
payment could be seen by a game in two moves and might be interpreted
as if CHI was paid twice by a carelessly implemented GSP.  We use the
last rather than first game, because this matches typical behaviour of
libraries that do not support duplicate keys explicitly; examples are
Python's json standard package and the JsonCpp library used in
libxayagame.

Note that this is a change compared to previous versions, where we
were only interpreting the first of duplicated fields.  Hence, updating
Xaya Core with this change forks game worlds (but not the Xaya network).
In practice, this is not an issue yet, as there are no important mainnet
game worlds out there at the moment.