Simplify and speed up name_expiration.py.
name_expiration.py was implemented before invalidateblock and
reconsiderblock were available to easily test reorgs.  Thus it is using
multiple nodes and a split / rejoined network.

Instead of that, we can now just use a single node and those RPCs to
perform the reorg orderly.  This simplifies the test (no need to keep
track of what nodes are on which chain), and also speeds it up massively
because there are no syncs between nodes involved anymore.