Simplify and speed up name_reorg.py.
Update name_reorg.py to use only a single node and the invalidateblock
and reconsiderblock RPCs to perform the reorg, rather than an actually
split network of four nodes.

This simplifies the test, as there are no longer multiple nodes with
different roles and states involved.  It also massively speeds it up, as
there is no longer any need to sync them.