Streamline name_registration.py.
This updates name_registration.py to mostly use only node 0, except
for things where we need two nodes.  This avoids most of the syncs,
speeding up the test massively.  It also simplifies the logic a bit,
since there are fewer instances where the exact index of the node
matters (since we mostly use node 0 for everything).