Add optional index for hash->name mapping.
This defines a new argument, -namehashindex.  When set, an indexer is run
that produces a database mapping hashes of registered names to the preimage
names themselves.

With this, we can (later on) support lookups of name by hash.  This then
allows asking an untrusted node for the existence of a name without
revealing the name itself if it does not yet exist.