Allow burning of coins with name operations.
This extends the "options" argument for name operations (name_register
and name_update), so that coins can easily be burnt in addition to the
existing "sendCoins".

To achieve this, a "burn" field should be passed.  The value must be a
JSON object.  The keys are put as "data" into the OP_RETURN outputs,
and the amounts are the number of coins that will be burnt in each output.

In a follow-up change, we will also pass data about burns to games through
the ZMQ interface.  This will then allow games to implement burns as
conditions for game actions (as an alternative to CHI payments).