Add regtest for atomic name trading.
This adds the new regtest name_ant_workflow.py, which performs
an atomic name trade between two wallets.  While there are already
some tests doing this in one way or another, the new test explicitly uses
the PSBT and raw-transaction interface in the way that is "recommended"
for doing ANTs in the most straight-forward way.  Thus it is useful
both as documentation for how to do it, and also to ensure this
method will not get broken and actually works.

(And in case we extend the interface in the future with easier-to-use
methods, then there is a place to "document" the new suggested workflow.)