Add NodeContext to WalletContext.
Unlike any upstream method, getauxblock is a wallet method but needs
the NodeContext.  This got broken with the upstream change in
https://github.com/bitcoin/bitcoin/pull/19096, because now the request
context is a WalletContext.

To fix it, we add a NodeContext pointer to WalletContext and in
this way pass both to the wallet methods.  EnsureNodeContext is able
to extract the NodeContext then accordingly in both cases.