Pass ChainstateManager to Namecoin code.
Properly pass down a ChainstateManager to the parts of the Namecoin
code that need it (mostly validation of the coin database and the
mempool, which is partly based on block heights).

This fixes the breakages introduced by the upstream removal of the
global BlockIndex in https://github.com/bitcoin/bitcoin/pull/19413.