Fix RPC client for namepsbt.
When namepsbt was added in https://github.com/namecoin/namecoin-core/pull/362,
the built-in RPC client was not updated and thus failed to correctly
convert the "vout" argument to a number and the "nameop" one to a
JSON object.  Hence, namecoin-cli and the built-in console in the Qt UI
were not able to use namepsbt.  This fixes the client accordingly.