Stats update script writing into MySQL.
This adds a new utility script for keeping track of Xaya stats
in a MySQL database (rather than just exporting to CSV).  The script
will connect to a database, query what blocks are in it, and then
attempt to fill in stats for all blocks that are missing compared to
current tip (minus 50 blocks to make room for reorgs).

So this script can e.g. be run with a cronjob to keep an almost-realtime
database of stats.