Implement bare hash for transactions.
This adds the concept of a bare hash ("btxid") for transactions.  This is
the hash of the transaction, but with all signature data removed (and
excluding any witnesses).  It matches the txid for unsigned transactions
or transactions with only witness data.

The bare hash commits to the "essential" data of a transaction, but is not
susceptible to malleability in any case, independent of whether or not the
inputs use segwit.  This makes it useful for tracking atomic trades.

The new btxid is returned from some RPC methods like decoderawtransaction,
and also sent as part of the ZMQ move data.